+++ itsl.rollup.js
@@ -5,57 +5,127 @@ const babel = require('rollup-plugin-babel');
 const { eslint } = require('rollup-plugin-eslint');
 const resolve = require('rollup-plugin-node-resolve');
 const scss = require('rollup-plugin-scss');
+const commonjs = require('rollup-plugin-commonjs');
 const svelte = require('rollup-plugin-svelte');
-const { uglify } = require('rollup-plugin-uglify');
+const { terser } = require('rollup-plugin-terser');
+
+const defaultOptions = {
+    legacy: false,
+    webComponents: false,
+    plugins: [],
+    eslint: {
+        configFile: 'node_modules/@itslearning/protomorph/.eslintrc.json',
+    },
+};
 
 /**
- * Returns a Rollup Configuration Object for Svelte files
+ * Returns a Rollup Configuration Object for Svelte files with optional polyfills and es5 compatibility
  * @param {string} src The source file
  * @param {string} dest The destination file
+ * @param {object} options
+ * @param {boolean} [options.legacy] Include polyfills and support for older browsers
+ * @param {boolean} [options.webComponents] Include polyfills for webComponents
+ * @param {any[]} [options.plugins] Array of plugins to run in addition to the defaults
+ * @param {object} [options.eslint] Eslint options, defaults to using protomorph eslintrc file
  * @returns {object} A Rollup Configuration Object
  */
-const Svelte = (src, dest, scriptPlugins = []) => ({
+const Svelte = (src, dest, options = defaultOptions) => ({
     input: src,
     output: {
         file: dest,
-        format: 'iife'
+        format: options.legacy ? 'iife' : 'esm',
+        sourcemap: !options.legacy,
     },
     treeshake: true,
     plugins: [
-        resolve({
-            extensions: ['.svelte', '.js']
+        options.legacy && prepareES5(src, options),
+        !options.legacy && eslint(options.eslint || defaultOptions.eslint),
+        // @ts-ignore
+        resolve({ extensions: [ '.js', '.mjs', '.html', '.svelte', '.json' ] }),
+        svelte({ extensions: ['.html', '.svelte'] }),
+        options.legacy && babelPreset,
+        // @ts-ignore
+        commonjs({
+            extensions: ['.js', '.mjs', '.html', '.svelte'],
+            namedExports: { 'chai': ['assert', 'expect'] },
         }),
-        eslint(),
-        svelte(),
-        babel({
-            babelrc: false,
-            presets: [['@babel/env', { modules: false }]],
-            extensions: ['.js', '.svelte']
-        }),
-        uglify(),
-        ...scriptPlugins
+        terser(),
+        ...options.plugins || defaultOptions.plugins
     ],
 });
+
+function prepareES5(src, options) {
+    const srcFile = path.resolve(src);
+
+    return {
+        name: 'Prepare bundle for ES5',
+        transform: (code, id) => {
+            if (srcFile !== id) {
+                return code;
+            } else {
+                return `
+import 'core-js/stable';
+import 'regenerator-runtime/runtime';
+import 'whatwg-fetch';
+${options.webComponents
+        ? "import '@webcomponents/webcomponentsjs/webcomponents-bundle.js';"
+        : ''
+}
+
+Promise.resolve(); // dummy call to trigger polyfill of Promise
+${code}`;
+            }
+        }
+    };
+}
+
+const babelPreset = babel({
+    exclude: [/\/core-js\//, '**/node_modules/@babel/runtime/**'],
+    babelrc: false,
+    externalHelpers: false,
+    runtimeHelpers: true,
+    presets: [
+        [
+            '@babel/preset-env',
+            {
+                useBuiltIns: 'entry',
+                corejs: 3,
+                targets: [ 'last 2 versions', 'not dead', 'ie 11' ],
+                modules: false,
+            }
+        ],
+    ],
+    extensions: [ '.js', '.mjs', '.html', '.svelte' ]
+});
+
+const sassOptions = {
+    plugins: [],
+};
 
 /**
  * Returns a Rollup Configuration Object for Scss files
  * @param {string} src The source file
  * @param {string} dest The destination file
+ * @param {object} options
+ * @param {any[]} options.plugins Array of plugins to run in addition to the defaults
  * @returns {object} A Rollup Configuration Object
  */
-const Sass = (src, dest, stylePlugins = []) => ({
+const Sass = (src, dest, options = sassOptions) => ({
     input: src,
     // Required for Rollup, just ignore
     output: {
         file: dest,
         format: 'esm'
     },
-    // Script will ALWAYS render an empty file at first
-    onwarn: (warning) => warning.code === 'EMPTY_BUNDLE' ? false : warning,
+    // Script will ALWAYS render an empty file at first, ignore EMPTY_BUNDLE
+    onwarn: (warning, onwarn) => warning.code === 'EMPTY_BUNDLE' || onwarn(warning),
     plugins: [
         scss({
             importer(path) {
-                return { file: path.replace(/^~/, 'node_modules/') };
+                return {
+                    file: path.replace(/^~/, 'node_modules/')
+                        .replace(/^@itslearning\//, 'node_modules/@itslearning/')
+                };
             },
             output: `${dest}.temp`,
             outputStyle: 'compact'
@@ -67,7 +137,7 @@ const Sass = (src, dest, stylePlugins = []) => ({
              */
             writeBundle: () => fs.renameSync(`${dest}.temp`, dest)
         },
-        ...stylePlugins
+        ...options.plugins || []
     ]
 });
 
@@ -75,29 +145,37 @@ const Sass = (src, dest, stylePlugins = []) => ({
  * Create an array of Rollup Configuration Objects
  * @param {object} config The required configuration
  * @param {string} config.destination The path where the generated file should be saved.
- * @param {Array<string>} config.files The files to be processed.
+ * @param {Array<string[]>} config.files The files to be processed.
+ * @param {any} config.plugins Additional rollup plugins to run.
  * @returns {object} A Rollup Configuration Object
  */
-const ItslRollup = ({ destination, files, plugins = {} }) =>
-    files.map(file => {
+const ItslRollup = ({ destination, files, plugins = {} }) => {
+    const configs = [];
+
+    files.forEach(file => {
         const inFile = Array.isArray(file) ? file[0] : file;
         const outFile = Array.isArray(file) ? file[1] || inFile : file;
 
-        const { ext } = path.parse(inFile);
+        const inPath = path.parse(inFile);
         const { name } = path.parse(outFile || inFile);
 
-        if (ext !== '.js' && ext !== '.scss') {
-            throw(`Unknown format ${ext}`);
+        if (inPath.ext !== '.js' && inPath.ext !== '.scss') {
+            throw (new Error(`Unknown format ${inPath.ext}`));
         }
-        return ext === '.js'
-            ? Svelte(inFile, `${destination}${name}.js`, plugins.script)
-            : ext === '.scss'
-            ? Sass(inFile, `${destination}${name}.css`, plugins.style)
-            : false;
+
+        if (inPath.ext === '.js') {
+            configs.push(Svelte(inFile, `${destination}${name}.js`, { legacy: false, plugins: plugins.script }));
+            configs.push(Svelte(inFile, `${destination}${name}.es5.js`, { legacy: true, plugins: plugins.script }));
+        } else if (inPath.ext === '.scss') {
+            configs.push(Sass(inFile, `${destination}${name}.css`, plugins.style));
+        }
     });
+
+    return configs;
+};
 
 module.exports = {
     ItslRollup,
     Svelte,
-    Sass
+    Sass,
 };